Support for multiple background images
This commit would give lozad.js the ability to support multiple background images as
<div style="background: url( 'path/to/main/image' ), url( 'path/to/fallback/image' );"></div>

The initialization will be not so different from existing way; simply add multiple paths/urls separated by commas:
<div class="lozad" data-background-image:"path/to/main/image,path/to/fallback/image"></div>

Then, lozad.js will split the data attribute by comma and join the parts again to match the CSS syntax.

This change is backward compatible; that is, it does not break existing code (of users of lozad.js).